requirements.txt
Flask
gunicorn
google-cloud-translate
